change method for installing alternc in the VM
if we compile the debian package first, then vagrant can use the .deb
file for installing alternc.
this is less dynamic for development, but the install process for
alternc is very strongly tied to packages for now.

I'm so close: preseed seems to mostly work but the alternc package fails
to install because /var/alternc doesn't exist.

also missing: management of the database